requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
numpy==2.2.2
packaging==24.2
PyOpenGL==3.1.9
pyopengltk==0.0.4
scipy==1.15.1
